+++ lion_master/src/app/modelling/rdf-models/vdi3682Model.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { PrefixesService } from '../../../shared/services/prefixes.service';
-import { QueriesService } from '../../../shared/services/backEnd/queries.service';
-import { GraphOperationsService } from '../../../shared/services/backEnd/graphOperations.service';
-import { DownloadService } from '../../../shared/services/backEnd/download.service';
-import { DataLoaderService } from '../../../shared/services/dataLoader.service';
-import { MessagesService } from '../../../shared/services/messages.service';
+import { PrefixesService } from '../../shared/services/prefixes.service';
+import { QueriesService } from '../../shared/services/backEnd/queries.service';
+import { GraphOperationsService } from '../../shared/services/backEnd/graphOperations.service';
+import { DownloadService } from '../../shared/services/backEnd/download.service';
+import { DataLoaderService } from '../../shared/services/dataLoader.service';
+import { MessagesService } from '../../shared/services/messages.service';
 import { take } from 'rxjs/operators';
 
 @Injectable({
