+++ lion_master/src/app/modelling/rdf-models/wadlModel.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { PrefixesService } from '../../../shared/services/prefixes.service';
-import { QueriesService } from '../../../shared/services/backEnd/queries.service';
-import { GraphOperationsService } from '../../../shared/services/backEnd/graphOperations.service';
-import { DataLoaderService } from '../../../shared/services/dataLoader.service';
-import { DownloadService } from '../../../shared/services/backEnd/download.service';
+import { PrefixesService } from '../../shared/services/prefixes.service';
+import { QueriesService } from '../../shared/services/backEnd/queries.service';
+import { GraphOperationsService } from '../../shared/services/backEnd/graphOperations.service';
+import { DataLoaderService } from '../../shared/services/dataLoader.service';
+import { DownloadService } from '../../shared/services/backEnd/download.service';
 import { Observable } from 'rxjs';
 import { take } from 'rxjs/operators';
